unsafe-path? changed to check path not parent of root-dir
serve-path now uses unsafe-path? to check the local-path formed by
combining root-dir and the selector.  This check doesn't raise an
exception if root-dir isn't safe or valid but it will return a path not
found error to the client.  This process should be safer and means that
we don't need to check root-dir every time serve-path is called at is
checked as part of the local-path check.